Add same-document links from headers to themselves
Callbacks are provided to allow control over how the links are
formatted.

The behaviour is guarded behind the MD_FLAG_HEADERSELFLINKS
parser flag.

We need to expose the MD_HTML type as an opaque type. The render self
link callback needs to know it to have the right type signatures.